# add _findall for AbstractVector{Bool} and use it in internal functions