--- a/src/other/utils.jl
+++ b/src/other/utils.jl
@@ -181,6 +181,34 @@ end
 
 _findall(B) = findall(B)
 
+function _findall(B::AbstractVector{Bool})
+    @assert firstindex(B) == 1
+    nnzB = count(B)
+
+    # fast path returning range
+    nnzB == 0 && return 1:0
+    nnzB == length(B) && return 1:length(B)
+    len = length(B)
+    start::Int = findfirst(B)
+    nnzB == 1 && return start:start
+    start + nnzB - 1 == len && return start:len
+    stop::Int = findnext(!, B, start+1) - 1
+    start + nnzB == stop + 1 && return start:stop
+
+    # slow path returning Vector{Int}
+    I = Vector{Int}(undef, nnzB)
+    I[1:stop - start + 1] .= start:stop
+    cnt = stop - start + 2
+    @inbounds for i in stop+1:len
+        if B[i]
+            I[cnt] = i
+            cnt += 1
+        end
+    end
+    @assert cnt == nnzB + 1
+    return I
+end
+
 @inline _blsr(x) = x & (x-1)
 
 # findall returning a range when possible (all true indices are contiguous), and optimized for B::BitVector
@@ -218,7 +246,7 @@ function _findall(B::BitVector)::Union{UnitRange{Int}, Vector{Int}}
         end
         if c == ~UInt64(0)
             if stop != -1
-                I = Vector{Int}(undef,nnzB)
+                I = Vector{Int}(undef, nnzB)
                 I[1:i-1] .= start:stop
                 break
             end