--- a/src/other/utils.jl
+++ b/src/other/utils.jl
@@ -246,7 +246,7 @@ function _findall(B::BitVector)::Union{UnitRange{Int}, Vector{Int}}
         end
         if c == ~UInt64(0)
             if stop != -1
-                I = Vector{Int}(undef, nnzB)
+                I = Vector{Int}(undef,nnzB)
                 I[1:i-1] .= start:stop
                 break
             end