Fixes https://github.com/JuliaData/DataFrames.jl/issues/2765