@bkamins 2021-05-21
Timings look acceptable:
```
julia> BD = OrderedDict(
           "T Big" => trues(100000),
           "F Big" => falses(100000),
           "T64 F64" => [trues(64); falses(64)],
           "F64 T64" => [falses(64); trues(64)],
           "F80 T100" => [falses(85); trues(100)],
           "F256 T32" => [falses(256); trues(32)],
           "F260 T32" => [falses(260); trues(32)],
           "TF Big" => [trues(100000); falses(100000)],
           "FT Big" => [falses(100000);trues(100000)] ,

           # some edge cases
           "TFT small" => [trues(85); falses(100); trues(85)],
           "FTFFT small" => [falses(64 + 32); trues(32); falses(128); trues(32)],
           "TFTF small" => [falses(64); trues(64); falses(64); trues(64)],
           "TFT small" => [trues(64); falses(10); trues(100)],

           "FTF Big" => [falses(8500); trues(100000); falses(65000)],
           "TFT Big" => [trues(8500); falses(100000); trues(65000)],
           "FTFTFTF Big" => [falses(65000); trues(65000); falses(65000); trues(65000); falses(65000); trues(65000); falses(65000)],

           "FTFR small" => [falses(85); trues(100); falses(65); rand([true, false], 20)],
           "R Big" => BitVector(rand([true, false], 200000)),
           "RF Big" => [BitVector(rand([true, false], 100000)) ; falses(100000)],
           "RT Big" => [BitVector(rand([true, false], 100000)) ; trues(100000)],
           "FTFR Big" => [falses(65000);  trues(65000);  falses(65000); rand([true, false], 20000)],
           "T256 R100" => [trues(256);  rand([true, false], 100)],
           "F256 R100" => [falses(256); rand([true, false], 100)],
       );

julia> for (l, B) in BD
           println("Processing $l")
           Bv = Vector{Bool}(B)
           @show Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv)
           @btime Base.findall($B)
           @btime DataFrames._findall($B)
           @btime Base.findall($Bv)
           @btime DataFrames._findall($Bv)
       end
Processing T Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  35.020 μs (2 allocations: 781.33 KiB)
  305.137 ns (0 allocations: 0 bytes)
  103.260 μs (2 allocations: 781.33 KiB)
  1.587 μs (0 allocations: 0 bytes)
Processing F Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  331.556 ns (1 allocation: 80 bytes)
  305.000 ns (0 allocations: 0 bytes)
  76.873 μs (1 allocation: 80 bytes)
  1.588 μs (0 allocations: 0 bytes)
Processing T64 F64
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  82.086 ns (1 allocation: 624 bytes)
  11.036 ns (0 allocations: 0 bytes)
  148.570 ns (1 allocation: 624 bytes)
  51.768 ns (0 allocations: 0 bytes)
Processing F64 T64
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  81.473 ns (1 allocation: 624 bytes)
  10.787 ns (0 allocations: 0 bytes)
  148.589 ns (1 allocation: 624 bytes)
  73.973 ns (0 allocations: 0 bytes)
Processing F80 T100
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  112.127 ns (1 allocation: 896 bytes)
  15.300 ns (0 allocations: 0 bytes)
  209.357 ns (1 allocation: 896 bytes)
  98.547 ns (0 allocations: 0 bytes)
Processing F256 T32
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  55.576 ns (1 allocation: 336 bytes)
  14.545 ns (0 allocations: 0 bytes)
  260.235 ns (1 allocation: 336 bytes)
  275.608 ns (0 allocations: 0 bytes)
Processing F260 T32
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  55.171 ns (1 allocation: 336 bytes)
  14.550 ns (0 allocations: 0 bytes)
  265.828 ns (1 allocation: 336 bytes)
  281.676 ns (0 allocations: 0 bytes)
Processing TF Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  79.269 μs (2 allocations: 781.33 KiB)
  1.399 μs (0 allocations: 0 bytes)
  180.256 μs (2 allocations: 781.33 KiB)
  53.601 μs (0 allocations: 0 bytes)
Processing FT Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  116.253 μs (2 allocations: 781.33 KiB)
  2.001 μs (0 allocations: 0 bytes)
  200.459 μs (2 allocations: 781.33 KiB)
  112.174 μs (0 allocations: 0 bytes)
Processing TFT small
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  176.293 ns (1 allocation: 1.45 KiB)
  138.606 ns (3 allocations: 1.53 KiB)
  233.015 ns (1 allocation: 1.45 KiB)
  192.385 ns (1 allocation: 1.45 KiB)
Processing FTFFT small
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  83.331 ns (1 allocation: 624 bytes)
  89.369 ns (3 allocations: 704 bytes)
  299.215 ns (1 allocation: 624 bytes)
  261.375 ns (1 allocation: 624 bytes)
Processing TFTF small
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  139.438 ns (1 allocation: 1.14 KiB)
  124.876 ns (5 allocations: 1.30 KiB)
  288.127 ns (1 allocation: 1.14 KiB)
  234.467 ns (1 allocation: 1.14 KiB)
Processing FTF Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  85.502 μs (2 allocations: 781.33 KiB)
  1.439 μs (0 allocations: 0 bytes)
  182.573 μs (2 allocations: 781.33 KiB)
  66.555 μs (0 allocations: 0 bytes)
Processing TFT Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  67.038 μs (2 allocations: 574.33 KiB)
  55.845 μs (2032 allocations: 653.62 KiB)
  170.520 μs (2 allocations: 574.33 KiB)
  126.508 μs (2 allocations: 574.33 KiB)
Processing FTFTFTF Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  176.434 μs (2 allocations: 1.49 MiB)
  129.894 μs (4062 allocations: 1.64 MiB)
  459.241 μs (2 allocations: 1.49 MiB)
  360.406 μs (2 allocations: 1.49 MiB)
Processing FTFR small
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  123.609 ns (1 allocation: 1008 bytes)
  97.650 ns (3 allocations: 1.06 KiB)
  307.919 ns (1 allocation: 1008 bytes)
  251.489 ns (1 allocation: 1008 bytes)
Processing R Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  108.669 μs (2 allocations: 780.33 KiB)
  85.125 μs (2 allocations: 780.33 KiB)
  913.936 μs (2 allocations: 780.33 KiB)
  771.246 μs (2 allocations: 780.33 KiB)
Processing RF Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  57.315 μs (2 allocations: 390.83 KiB)
  42.894 μs (2 allocations: 390.83 KiB)
  563.630 μs (2 allocations: 390.83 KiB)
  439.797 μs (2 allocations: 390.83 KiB)
Processing RT Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  149.267 μs (2 allocations: 1.14 MiB)
  113.986 μs (3126 allocations: 1.26 MiB)
  564.277 μs (2 allocations: 1.14 MiB)
  460.027 μs (2 allocations: 1.14 MiB)
Processing FTFR Big
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  70.646 μs (2 allocations: 586.58 KiB)
  44.032 μs (4 allocations: 586.66 KiB)
  273.804 μs (2 allocations: 586.58 KiB)
  252.519 μs (2 allocations: 586.58 KiB)
Processing T256 R100
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  446.406 ns (1 allocation: 2.62 KiB)
  364.292 ns (3 allocations: 2.70 KiB)
  581.854 ns (1 allocation: 2.62 KiB)
  487.621 ns (1 allocation: 2.62 KiB)
Processing F256 R100
Base.findall(B) == DataFrames._findall(B) == DataFrames._findall(Bv) = true
  76.712 ns (1 allocation: 544 bytes)
  69.048 ns (1 allocation: 544 bytes)
  332.305 ns (1 allocation: 544 bytes)
  381.655 ns (1 allocation: 544 bytes)
```