@pstorozenko 2021-05-22
Nice, I had not time to look on it during a week.
I wanted to suggest checking, where `_findall` could be used internally, good that you looked on it.
Maybe add this information to the pull request name?