@bkamins 2021-05-22
Tests of things changed (all timings after compilation; I mostly focus on checking if we do not regress).
In general things look good (except one case on which I comment).

This PR:
```
julia> f(row) = row.x > 0.5; df = DataFrame(x = 0:0.0000001:1);

julia> @time filter!(AsTable(:x) => f, df);
  0.055787 seconds (42 allocations: 77.493 MiB)

julia> df = DataFrame(x = rand(10^6));

julia> @time filter!(AsTable(:x) => f, df);
  0.010720 seconds (43 allocations: 11.571 MiB)

julia> @time unique!(df);
  0.027130 seconds (20 allocations: 12.290 MiB)

julia> df = DataFrame(x = rand(1:10^4, 10^6));

julia> @time unique!(df);
  0.025073 seconds (41 allocations: 32.137 MiB, 5.88% gc time)

julia> df = DataFrame(rand(4, 1000), :auto);

julia> @time df[[true, true, true, true], 1:end]; # this and below is the only regression
  0.000262 seconds (2.99 k allocations: 206.812 KiB)

julia> @time df[[true, true, true, true], :]; # this is the second regression - I comment on it in the code
  0.000244 seconds (2.99 k allocations: 206.469 KiB)

julia> df = DataFrame(rand(2*10^6, 64), :auto);

julia> @time df[trues(nrow(df)), 1:end];
  0.178173 seconds (674 allocations: 976.846 MiB)

julia> @time df[trues(nrow(df)), :];
  0.168878 seconds (670 allocations: 976.845 MiB)

julia> df = DataFrame(x = rand(10^6));

julia> @time delete!(df, trues(10^6));
  0.000069 seconds (7 allocations: 122.328 KiB)

julia> df1 = DataFrame(x = 1:10^6);

julia> df2 = DataFrame(x = 10^5:10^6+10^5);

julia> @btime outerjoin($df1, $df2, on=:x);
  20.150 ms (259 allocations: 23.91 MiB)

julia> @btime outerjoin($df1, $df2, on=:x);
  20.352 ms (259 allocations: 23.91 MiB)
```

`main`:
```
julia> f(row) = row.x > 0.5; df = DataFrame(x = 0:0.0000001:1);

julia> @time filter!(AsTable(:x) => f, df);
  0.093865 seconds (48 allocations: 115.640 MiB, 7.79% gc time)

julia> df = DataFrame(x = rand(10^6));

julia> @time filter!(AsTable(:x) => f, df);
  0.012052 seconds (48 allocations: 11.567 MiB)

julia> @time unique!(df);
  0.023995 seconds (18 allocations: 12.296 MiB)

julia> df = DataFrame(x = rand(1:10^4, 10^6));

julia> @time unique!(df);
  0.025348 seconds (39 allocations: 32.137 MiB)

julia> df = DataFrame(rand(4, 1000), :auto);

julia> @time df[[true, true, true, true], 1:end];
  0.000268 seconds (1.99 k allocations: 175.672 KiB)

julia> @time df[[true, true, true, true], :];
  0.000250 seconds (1.99 k allocations: 175.328 KiB)

julia> df = DataFrame(rand(2*10^6, 64), :auto);

julia> @time df[trues(nrow(df)), 1:end];
  0.195522 seconds (614 allocations: 992.104 MiB)

julia> @time df[trues(nrow(df)), :];
  0.180493 seconds (609 allocations: 992.102 MiB)

julia> df = DataFrame(x = rand(10^6));

julia> @time delete!(df, trues(10^6));
  0.002477 seconds (8 allocations: 7.749 MiB)

julia> df1 = DataFrame(x = 1:10^6);

julia> df2 = DataFrame(x = 10^5:10^6+10^5);

julia> @btime outerjoin($df1, $df2, on=:x);
  20.936 ms (270 allocations: 25.43 MiB)

julia> @btime outerjoin($df1, $df2, on=:x);
  20.892 ms (270 allocations: 25.43 MiB)
```